refactor: extract api handlers and align tests
