+++ src/app/api/spreadsheet/health/health-handler.ts
@@ -1,0 +1,39 @@
+// ABOUTME: Shared handler factory for spreadsheet health API.
+// ABOUTME: Allows tests to inject diagnostic fetchers while keeping route exports minimal.
+import { NextResponse } from "next/server";
+
+import { fetchSpreadsheetDiagnostics } from "@/server/google/spreadsheet-health";
+
+type FetchDiagnostics = typeof fetchSpreadsheetDiagnostics;
+
+function isUnauthorized(message: string) {
+  return message === "Missing authenticated session" || message === "Missing Google tokens";
+}
+
+export function createSpreadsheetHealthHandler({
+  fetchDiagnostics = fetchSpreadsheetDiagnostics,
+}: {
+  fetchDiagnostics?: FetchDiagnostics;
+} = {}) {
+  const GET = async (request: Request) => {
+    const url = new URL(request.url);
+    const spreadsheetId = url.searchParams.get("spreadsheetId")?.trim() ?? "";
+
+    if (!spreadsheetId) {
+      return NextResponse.json({ error: "Missing spreadsheetId" }, { status: 400 });
+    }
+
+    try {
+      const diagnostics = await fetchDiagnostics({ spreadsheetId });
+
+      return NextResponse.json({ diagnostics }, { status: 200 });
+    } catch (error) {
+      const message = error instanceof Error ? error.message : "Unknown error";
+      const status = isUnauthorized(message) ? 401 : 500;
+
+      return NextResponse.json({ error: message }, { status });
+    }
+  };
+
+  return { GET };
+}
