+++ src/server/google/repository/accounts-repository.ts
@@ -284,7 +284,7 @@ export function createAccountsRepository({
           spreadsheetId,
           range,
           valueInputOption: "RAW",
-          resource: {
+          requestBody: {
             values: rows,
           },
         }),
