+++ src/server/google/repository/actuals-repository.ts
@@ -146,7 +146,7 @@ export function createActualsRepository({
           spreadsheetId,
           range,
           valueInputOption: "RAW",
-          resource: {
+          requestBody: {
             values: rows,
           },
         }),
