+++ src/server/google/repository/future-events-repository.ts
@@ -158,7 +158,7 @@ export function createFutureEventsRepository({
           spreadsheetId,
           range,
           valueInputOption: "RAW",
-          resource: {
+          requestBody: {
             values: rows,
           },
         }),
